Fix build blockers: Liquibase config, missing migrations, entity and repo bugs

+++ LocalMarketplace/src/main/java/bg/sofia/uni/fmi/localmarketplace/repository/cart/CartItemRepository.java
@@ -3,5 +3,5 @@ package bg.sofia.uni.fmi.localmarketplace.repository.cart;
 import bg.sofia.uni.fmi.localmarketplace.domain.cart.CartItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
+public interface CartItemRepository extends JpaRepository<CartItem, Long> {
 }
